pyproject.toml
[project]
name = "ducklake-scripts"
version = "0.1.0"
description = "Scripts for managing SSL certificates and GitHub secrets"
requires-python = ">=3.8"
dependencies = [
    "requests>=2.25.0",
    "pynacl>=1.4.0",
    "snowflake>=1.0.2",
    "pyarrow>=17.0.0",
    "adbc-driver-flightsql>=1.2.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["scripts"]

[tool.uv]
dev-dependencies = []
